refactor: input base
