+++ src/components/Input/inputBase/styled/label.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+import { COLOR_GRAY_4 } from '../../../../styles/colors';
+import { FONT_SIZE_TEXT_MEDIUM } from '../../../../styles/fontSizes';
+
+const Label = styled.label`
+    color: ${COLOR_GRAY_4};
+    font-size: ${FONT_SIZE_TEXT_MEDIUM};
+    line-height: 1.5;
+    margin-bottom: 0.125rem;
+    align-self: center;
+
+    :empty {
+        margin: 0;
+    }
+
+    ${props => props.readOnly && !props.disabled && 'align-self: flex-start;'};
+`;
+
+export default Label;
